(Refactor) move CompanyCard and MapOverlay components to entities folder & update imports
