+++ src/entities/map/ui/CompanyCard.tsx
